made changes to comments

+++ libraries/ms-common/src/stm32f0xx/chip_id.c
@@ -6,13 +6,13 @@
 // datasheet
 #define DEVICE_ID_ADDR 0x1FFFF7AC
 
-// returns chip id in the form of an array of uint16_ts
+// returns chip id in the form of an array of uint32_ts
 ChipId chip_id_get(void) {
   ChipId chip_id = { 0 };
-  uint16_t *address_pointer = (uint16_t *)DEVICE_ID_ADDR;
+  uint32_t *address_pointer = (uint32_t *)DEVICE_ID_ADDR;
 
   // Iterates through 96 bits of id and stores them in uint32_t array
-  for (int i = 0; i < 2; i++) {
+  for (uint8_t i = 0; i < 2; i++) {
     chip_id.id[i] = *address_pointer;
 
     // Offsets address by 0x04
